refactor: finish action/button schema refactor
This concludes the LED state refactoring project. The buttons have been
updated to simply trigger Actions, which aren't tied to any specific
button. One of the action options is setting LEDs for buttons, which is
how the LED state management is achieved now. Each button bind also has
a default state setting which is how the LEDs are initialized.